arm-cm3: context - primitives for threads and task switching
This will definitely change -- not entirely happy with it as is -- but it has
some useful bits for building threads on top of.

context_init(entry,new_psp,new_msp) sets up the PSP and MSP stacks, switches
to PSP, and jumps to entry.

handle_m3_pendsv() looks at the global_context state (4 words at the top
of the MSP) and does a simple context switch.